Feat: 프로필 조회에 게시글 좋아요 수와 토론 참여자 수 추가
DTO 기존 Integer에서 Long으로 변경

+++ src/main/java/com/example/mssaem_backend/domain/boardcomment/BoardCommentRepository.java
@@ -17,7 +17,7 @@ public interface BoardCommentRepository extends JpaRepository<BoardComment, Long
 
     List<BoardComment> findAllByBoardId(Long id);
 
-    Integer countAllByStateIsTrueAndMember(@Param("member") Member member);
+    Long countAllByStateIsTrueAndMember(@Param("member") Member member);
 
     @Query(value = "select bc from BoardComment bc join fetch bc.member where bc.board.id = :id", countQuery = "select count(bc) from BoardComment bc")
     Page<BoardComment> findAllByBoardId(@Param("id") Long id, Pageable pageable);
